move demo endpoint to fast-endpoint
