+++ MyNetatmo24.Backend/MyNetatmo24.Backend.csproj
@@ -8,6 +8,8 @@
 
     <ItemGroup>
         <PackageReference Include="FastEndpoints" Version="5.30.0" />
+        <PackageReference Include="FastEndpoints.Security" Version="5.30.0" />
+        <PackageReference Include="FastEndpoints.Swagger" Version="5.30.0" />
         <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="8.0.8" />
         <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="8.0.7"/>
         <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0"/>
